# Codigos/Back-End/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API - Sistema de Controle de Fluxos Veiculares",
    "description": "API para gerenciamento de usuários, integradores, municípios, equipamentos e fluxos veiculares.",
    "version": "1.0.0",
    "contact": {
      "name": "DNIT",
      "url": "https://www.gov.br/dnit"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Servidor de desenvolvimento"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Integrador": {
        "type": "object",
        "required": ["nome"],
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "nome": {
            "type": "string",
            "maxLength": 255,
            "example": "Integrador XYZ"
          }
        }
      },
...
      "Fluxo": {
        "type": "object",
        "required": [
          "seq",
          "data",
          "hora",
          "classVeic",
          "dataRecebimento",
          "equipamentoId"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "seq": {
            "type": "integer",
            "example": 12345
          },
          "data": {
            "type": "string",
            "format": "date",
            "example": "2025-07-21"
          },
          "hora": {
            "type": "string",
            "format": "time",
            "example": "14:30:00"
          },
          "placa": {
            "type": "string",
            "maxLength": 7,
            "example": "ABC1D23"
          },
          "velMed": {
            "type": "integer",
            "minimum": 0,
            "example": 85
          },
          "tamVeic": {
            "type": "integer",
            "minimum": 0,
            "example": 420
          },
          "classVeic": {
            "type": "string",
            "maxLength": 3,
            "example": "3"
          },
          "pesoBt": {
            "type": "integer",
            "minimum": 0,
            "example": 3500
          },
          "dataRecebimento": {
            "type": "string",
            "format": "date-time",
            "example": "2025-07-21T14:30:45Z"
          },
          "equipamentoId": {
            "type": "integer",
            "example": 1
          }
        }
      },
...
    },
    "parameters": {
      "idPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer"
        },
        "description": "ID do recurso"
      },
      "limitQuery": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 30,
          "maximum": 100
        },
        "description": "Limite de resultados retornados"
      }
    }
  },
  "paths": {
...
    "/integradores": {
      "get": {
        "tags": ["Integradores"],
        "summary": "Listar integradores",
        "responses": {
          "200": {
            "description": "Lista de integradores",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Integrador"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Integradores"],
        "summary": "Criar integrador",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Integrador"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Integrador criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Integrador"
                }
              }
            }
          }
        }
      }
    },
    "/integradores/{id}": {
      "get": {
        "tags": ["Integradores"],
        "summary": "Obter integrador por ID",
        "parameters": [
          {
            "$ref": "#/components/parameters/idPath"
          }
        ],
        "responses": {
          "200": {
            "description": "Integrador encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Integrador"
                }
              }
            }
          },
          "404": {
            "description": "Integrador não encontrado"
          }
        }
      },
      "put": {
        "tags": ["Integradores"],
        "summary": "Atualizar integrador",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "$ref": "#/components/parameters/idPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Integrador"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Integrador atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Integrador"
                }
              }
            }
          },
          "404": {
            "description": "Integrador não encontrado"
          }
        }
      },
      "delete": {
        "tags": ["Integradores"],
        "summary": "Excluir integrador",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "$ref": "#/components/parameters/idPath"
          }
        ],
        "responses": {
          "204": {
            "description": "Integrador excluído com sucesso"
          },
          "404": {
            "description": "Integrador não encontrado"
          }
        }
      }
    },
...
    "/fluxos": {
      "get": {
        "tags": ["Fluxos"],
        "summary": "Listar fluxos",
        "security": [{ "bearerAuth": [] }],
        "description": "Retorna os registros de fluxo veicular com filtros opcionais. Limite máximo de 30 registros por padrão.",
        "parameters": [
          {
            "name": "CodEquipamento",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Código do equipamento"
          },
          {
            "name": "faixaEquipamento",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Faixa do equipamento"
          },
          {
            "name": "dataInicio",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Data de início (YYYY-MM-DD)"
          },
          {
            "name": "dataFim",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Data de fim (YYYY-MM-DD)"
          },
          {
            "name": "horaInicio",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^([0-1][0-9]|2[0-3]):[0-5][0-9]$"
            },
            "description": "Hora de início (HH:MM)"
          },
          {
            "name": "horaFim",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^([0-1][0-9]|2[0-3]):[0-5][0-9]$"
            },
            "description": "Hora de fim (HH:MM)"
          },
          {
            "name": "placa",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Placa do veículo"
          },
          {
            "$ref": "#/components/parameters/limitQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de fluxos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Fluxo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/fluxos/{id}": {
      "get": {
        "tags": ["Fluxos"],
        "summary": "Obter fluxo",
        "parameters": [
          {
            "$ref": "#/components/parameters/idPath"
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Fluxo encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fluxo"
                }
              }
            }
          },
          "404": {
            "description": "Fluxos não encontrado"
          }
        }
      },
      "delete": {
        "tags": ["Fluxos"],
        "summary": "Excluir fluxo",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "$ref": "#/components/parameters/idPath"
          }
        ],
        "responses": {
          "204": {
            "description": "Fluxo excluído com sucesso"
          },
          "404": {
            "description": "Fluxo não encontrado"
          }
        }
      }
    },
    "/fluxos/count/{equipamento_id}": {
      "get": {
        "tags": ["Fluxos"],
        "summary": "Obter a quantidade de fluxos por equipamento",
        "parameters": [
          {
            "name": "equipamento_id",
            "in": "path",
            "schema": {
              "type": "number",
              "required": true
            },
            "description": "ID do equipamento"
          },
          {
            "name": "dataInicio",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Data de início (YYYY-MM-DD)"
          },
          {
            "name": "dataFim",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Data de fim (YYYY-MM-DD)"
          },
          {
            "name": "horaInicio",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^([0-1][0-9]|2[0-3]):[0-5][0-9]$"
            },
            "description": "Hora de início (HH:MM)"
          },
          {
            "name": "horaFim",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^([0-1][0-9]|2[0-3]):[0-5][0-9]$"
            },
            "description": "Hora de fim (HH:MM)"
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Contagem de Fluxos por equipamento encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "exemple": 1
                }
              }
            }
          },
          "404": {
            "description": "Contagem de Fluxos por equipamento não encontrada"
          }
        }
      }
    }
  }
}